Extracted SHA identity hashing into SearchIdentity
* Split SearchParameterHash/LibraryIdentityHash/ReconciliationParameterHash + EscapeForRustDebug out of the OspreyConfig mutable bag into a single-responsibility SearchIdentity
* OspreyConfig keeps thin delegators + an Identity property; recipes moved verbatim, byte-identical (CoreTypesTest hash oracle green)
* First step of PR-A in the declarative-dataflow program

See ai/todos/active/TODO-20260601_ospreysharp_declarative_pipeline_dataflow.md
