+++ pwiz_tools/OspreySharp/OspreySharp.Core/OspreyConfig.cs
@@ -23,9 +23,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Security.Cryptography;
-using System.Text;
 
 namespace pwiz.OspreySharp.Core
 {
@@ -208,267 +205,44 @@ namespace pwiz.OspreySharp.Core
         }
 
         /// <summary>
-        /// Compute SHA-256 hash of parameters that affect first-pass scoring.
-        /// If this hash changes, cached .scores.parquet files are invalid.
+        /// The bit-parity-critical identity hashing for this run. Split out
+        /// of <see cref="OspreyConfig"/> into <see cref="Core.SearchIdentity"/>
+        /// so this type is only the configuration bag and the SHA hashing is
+        /// its own single-responsibility unit. A fresh instance is returned
+        /// per access; it reads this config's hash-affecting fields at call
+        /// time, preserving the historical behavior of the former instance
+        /// methods. The hash recipes live on <see cref="Core.SearchIdentity"/>
+        /// and MUST stay byte-identical with Rust.
         /// </summary>
-        public string SearchParameterHash()
-        {
-            // Cross-impl bit-equivalence with Rust requires:
-            //  - Booleans: Rust prints "true"/"false" (lowercase). C# default
-            //    bool.ToString() is "True"/"False". Use lowercase explicitly.
-            //  - Numbers: invariant culture (no locale-dependent separators).
-            using (var sha256 = SHA256.Create())
-            {
-                var ic = System.Globalization.CultureInfo.InvariantCulture;
-                Func<bool, string> b = v => v ? "true" : "false";
-                var sb = new StringBuilder();
-                sb.AppendFormat(ic, "resolution_mode:{0}\n", ResolutionMode);
-                sb.AppendFormat(ic, "fragment_tolerance:{0},{1}\n", FragmentTolerance.Tolerance, FragmentTolerance.Unit);
-                sb.AppendFormat(ic, "precursor_tolerance:{0},{1}\n", PrecursorTolerance.Tolerance, PrecursorTolerance.Unit);
-                sb.AppendFormat(ic, "prefilter_enabled:{0}\n", b(PrefilterEnabled));
-                sb.AppendFormat(ic, "decoy_method:{0}\n", DecoyMethod);
-                sb.AppendFormat(ic, "decoys_in_library:{0}\n", b(DecoysInLibrary));
-                // Sort prefixes so ordering changes don't churn the hash.
-                // Lower-case to make case-only edits no-ops (matching the
-                // runtime comparison). Mirrors Rust's
-                // format!("decoy_prefixes:{:?}\n", prefixes) where {:?} on
-                // Vec<String> yields ["a", "b"] (double-quoted, comma-
-                // space-separated).
-                var prefixes = new List<string>(DecoyPrefixes != null ? DecoyPrefixes.Count : 0);
-                if (DecoyPrefixes != null)
-                {
-                    foreach (var p in DecoyPrefixes)
-                        prefixes.Add(p == null ? string.Empty : p.ToLowerInvariant());
-                }
-                prefixes.Sort(StringComparer.Ordinal);
-                var prefixList = new StringBuilder("[");
-                for (int i = 0; i < prefixes.Count; i++)
-                {
-                    if (i > 0) prefixList.Append(", ");
-                    prefixList.Append('"').Append(prefixes[i]).Append('"');
-                }
-                prefixList.Append(']');
-                sb.AppendFormat(ic, "decoy_prefixes:{0}\n", prefixList.ToString());
-                // Mirror Rust's `format!("decoy_pairing_manifest:{:?}\n", ...)`
-                // where the value is `Some("path")` or `None`. Rust's {:?}
-                // on String escapes \, ", \n, \r, \t, \0, and other control
-                // chars (< 0x20) -- critical for Windows paths whose `\`
-                // separators must become `\\` to match Rust's output. The
-                // path is not normalised; the user's choice (relative or
-                // absolute) is part of the hash so a moved manifest
-                // invalidates the cache.
-                sb.AppendFormat(ic, "decoy_pairing_manifest:{0}\n",
-                    string.IsNullOrEmpty(DecoyPairingManifestPath)
-                        ? "None"
-                        : "Some(\"" + EscapeForRustDebug(DecoyPairingManifestPath) + "\")");
-                sb.AppendFormat(ic, "decoy_pair_min_fraction:{0}\n", DecoyPairMinFraction);
-                sb.AppendFormat(ic, "rt_cal.enabled:{0}\n", b(RtCalibration.Enabled));
-                sb.AppendFormat(ic, "rt_cal.fallback_rt_tolerance:{0}\n", RtCalibration.FallbackRtTolerance);
-                sb.AppendFormat(ic, "rt_cal.rt_tolerance_factor:{0}\n", RtCalibration.RtToleranceFactor);
-                sb.AppendFormat(ic, "rt_cal.min_rt_tolerance:{0}\n", RtCalibration.MinRtTolerance);
-                sb.AppendFormat(ic, "rt_cal.max_rt_tolerance:{0}\n", RtCalibration.MaxRtTolerance);
-                sb.AppendFormat(ic, "rt_cal.loess_bandwidth:{0}\n", RtCalibration.LoessBandwidth);
-                sb.AppendFormat(ic, "rt_cal.min_calibration_points:{0}\n", RtCalibration.MinCalibrationPoints);
-                sb.AppendFormat(ic, "rt_cal.calibration_sample_size:{0}\n", RtCalibration.CalibrationSampleSize);
-                sb.AppendFormat(ic, "rt_cal.calibration_retry_factor:{0}\n", RtCalibration.CalibrationRetryFactor);
-                sb.AppendFormat(ic, "reconciliation.top_n_peaks:{0}\n", Reconciliation.TopNPeaks);
-
-                byte[] hashBytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(sb.ToString()));
-                var result = new StringBuilder(64);
-                for (int i = 0; i < hashBytes.Length; i++)
-                {
-                    result.Append(hashBytes[i].ToString("x2"));
-                }
-                return result.ToString();
-            }
-        }
+        public SearchIdentity Identity => new SearchIdentity(this);
 
         /// <summary>
-        /// Compute a fast identity hash for the library file (file name + size
-        /// + mtime). Filesystem metadata only -- no content hashing. The
-        /// directory portion is deliberately NOT in the hash so the same
-        /// library identifies identically across Rust / .NET / OS variations
-        /// (drive letter case, forward vs back slash, relative vs absolute,
-        /// HPC node-local vs shared paths). Mirrors the
-        /// <c>reconciliation_parameter_hash</c> precedent that hashes only
-        /// sorted file stems for the input set. Same recipe as Rust's
-        /// <c>library_identity_hash</c>.
+        /// Convenience delegator to <see cref="Core.SearchIdentity.SearchParameterHash"/>.
         /// </summary>
-        public string LibraryIdentityHash()
-        {
-            string libPath = LibrarySource != null ? LibrarySource.Path : string.Empty;
-            using (var sha256 = SHA256.Create())
-            {
-                var sb = new StringBuilder();
-                string fileName = string.IsNullOrEmpty(libPath)
-                    ? string.Empty
-                    : Path.GetFileName(libPath);
-                sb.AppendFormat("file_name:{0}\n", fileName);
-                if (!string.IsNullOrEmpty(libPath) && File.Exists(libPath))
-                {
-                    var info = new FileInfo(libPath);
-                    sb.AppendFormat(System.Globalization.CultureInfo.InvariantCulture,
-                        "size:{0}\n", info.Length);
-                    // Unix seconds matching Rust's library_identity_hash
-                    // (SystemTime::duration_since(UNIX_EPOCH).as_secs()).
-                    long mtimeSecs = (long)(info.LastWriteTimeUtc
-                        - new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc)).TotalSeconds;
-                    sb.AppendFormat(System.Globalization.CultureInfo.InvariantCulture,
-                        "mtime:{0}\n", mtimeSecs);
-                }
-                byte[] hashBytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(sb.ToString()));
-                var result = new StringBuilder(64);
-                for (int i = 0; i < hashBytes.Length; i++)
-                    result.Append(hashBytes[i].ToString("x2"));
-                return result.ToString();
-            }
-        }
+        public string SearchParameterHash() => Identity.SearchParameterHash();
 
         /// <summary>
-        /// Escape a string to match Rust's <c>{:?}</c> Debug formatter
-        /// output for <c>&amp;str</c> / <c>String</c>. Handles the cases
-        /// that actually appear in config values folded into the search
-        /// hash: backslashes, double quotes, common C escapes
-        /// (<c>\n</c>, <c>\r</c>, <c>\t</c>, <c>\0</c>), and other
-        /// sub-0x20 control characters via the <c>\u{...}</c> form.
-        /// Printable ASCII and non-ASCII bytes pass through unchanged --
-        /// matching Rust's default Debug output as of the language
-        /// versions in use by maccoss/osprey (1.75+). Used so cross-impl
-        /// hashes agree on Windows paths whose <c>\</c> separators must
-        /// render as <c>\\</c>.
+        /// Convenience delegator to <see cref="Core.SearchIdentity.LibraryIdentityHash"/>.
         /// </summary>
-        internal static string EscapeForRustDebug(string s)
-        {
-            if (string.IsNullOrEmpty(s))
-                return s ?? string.Empty;
-            var sb = new StringBuilder(s.Length + 8);
-            foreach (char c in s)
-            {
-                switch (c)
-                {
-                    case '\\': sb.Append(@"\\"); break;
-                    case '"':  sb.Append("\\\""); break;
-                    case '\n': sb.Append(@"\n"); break;
-                    case '\r': sb.Append(@"\r"); break;
-                    case '\t': sb.Append(@"\t"); break;
-                    case '\0': sb.Append(@"\0"); break;
-                    default:
-                        if (c < 0x20 || c == 0x7F)
-                        {
-                            // Rust's {:?} uses `\u{HEX}` with lowercase hex,
-                            // no padding. AppendFormat's `{0:x}` works on
-                            // net8.0 but produces `{x}` literally on net472
-                            // under the double-brace escape sequence; use
-                            // explicit ToString to avoid the regression.
-                            sb.Append(@"\u{");
-                            sb.Append(((int)c).ToString(@"x",
-                                System.Globalization.CultureInfo.InvariantCulture));
-                            sb.Append('}');
-                        }
-                        else
-                        {
-                            sb.Append(c);
-                        }
-                        break;
-                }
-            }
-            return sb.ToString();
-        }
+        public string LibraryIdentityHash() => Identity.LibraryIdentityHash();
 
         /// <summary>
-        /// SHA-256 of (search hash + reconciliation parameters + run FDR
-        /// + sorted file stems). Mirrors Rust
-        /// <c>OspreyConfig::reconciliation_parameter_hash</c> in
-        /// <c>crates/osprey-core/src/config.rs</c> and is written into
-        /// reconciled <c>.scores.parquet</c> footer metadata under
-        /// <c>osprey.reconciliation_hash</c>. The hash invalidates the
-        /// cache on any reconciliation parameter change OR on any change
-        /// to the multi-file set (file_stems are sorted to make the
-        /// hash invariant to invocation order).
+        /// Convenience delegator to <see cref="Core.SearchIdentity.ReconciliationParameterHash"/>.
         /// </summary>
-        public string ReconciliationParameterHash()
-        {
-            var stems = new List<string>(InputFiles?.Count ?? 0);
-            if (InputFiles != null)
-            {
-                foreach (var path in InputFiles)
-                {
-                    string stem = Path.GetFileNameWithoutExtension(path);
-                    if (!string.IsNullOrEmpty(stem))
-                        stems.Add(stem);
-                }
-            }
-            return ReconciliationParameterHashForStems(stems);
-        }
+        public string ReconciliationParameterHash() => Identity.ReconciliationParameterHash();
 
         /// <summary>
-        /// Compute the reconciliation parameter hash for an explicit set of
-        /// file stems. Used by per-file Stage 6 rescore workers, whose
-        /// <see cref="InputFiles"/> only carries this worker's single
-        /// parquet — the hash that the downstream <c>--join-at-pass=2</c>
-        /// merge node expects is computed over ALL files in the join, so
-        /// the worker must read the full set from the planner's
-        /// <c>reconciliation.json</c> envelope and pass it in here. The
-        /// stems are sorted + deduped internally so the hash is invariant
-        /// to caller ordering. Mirrors Rust
-        /// <c>OspreyConfig::reconciliation_parameter_hash_for_stems</c>.
+        /// Convenience delegator to <see cref="Core.SearchIdentity.ReconciliationParameterHashForStems"/>.
         /// </summary>
         public string ReconciliationParameterHashForStems(IReadOnlyList<string> fileStems)
-        {
-            using (var sha256 = SHA256.Create())
-            {
-                var ic = System.Globalization.CultureInfo.InvariantCulture;
-                var sb = new StringBuilder();
-                sb.Append(SearchParameterHash());
-                sb.AppendFormat(ic, "reconciliation.enabled:{0}\n",
-                    Reconciliation.Enabled ? "true" : "false");
-                sb.AppendFormat(ic, "reconciliation.consensus_fdr:{0}\n",
-                    Reconciliation.ConsensusFdr);
-                sb.AppendFormat(ic, "run_fdr:{0}\n", RunFdr);
-                // Mirror Rust's `format!("file_stems:{:?}\n", stems)` output
-                // exactly. {:?} on Vec<String> yields ["a", "b"] with the
-                // brackets and double-quoted, comma-space-separated values.
-                // Stems are sorted + deduped here so the hash matches the
-                // Rust side, which also sorts + dedups before hashing.
-                var stems = new List<string>(fileStems?.Count ?? 0);
-                if (fileStems != null)
-                {
-                    foreach (var stem in fileStems)
-                    {
-                        if (!string.IsNullOrEmpty(stem))
-                            stems.Add(stem);
-                    }
-                }
-                stems.Sort(StringComparer.Ordinal);
-                // Dedup in place (stems is sorted, so duplicates are
-                // adjacent). Rust does `dedup()` on a sorted Vec; same here.
-                int write = 0;
-                for (int read = 0; read < stems.Count; read++)
-                {
-                    if (read == 0 || !string.Equals(stems[read], stems[read - 1], StringComparison.Ordinal))
-                    {
-                        stems[write++] = stems[read];
-                    }
-                }
-                if (write < stems.Count)
-                    stems.RemoveRange(write, stems.Count - write);
+            => Identity.ReconciliationParameterHashForStems(fileStems);
 
-                var stemsList = new StringBuilder("[");
-                for (int i = 0; i < stems.Count; i++)
-                {
-                    if (i > 0) stemsList.Append(", ");
-                    stemsList.Append('"').Append(stems[i]).Append('"');
-                }
-                stemsList.Append(']');
-                sb.AppendFormat(ic, "file_stems:{0}\n", stemsList.ToString());
-
-                byte[] hashBytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(sb.ToString()));
-                var result = new StringBuilder(64);
-                for (int i = 0; i < hashBytes.Length; i++)
-                    result.Append(hashBytes[i].ToString("x2"));
-                return result.ToString();
-            }
-        }
+        /// <summary>
+        /// Convenience delegator to <see cref="Core.SearchIdentity.EscapeForRustDebug"/>,
+        /// retained so existing callers / tests that referenced
+        /// <c>OspreyConfig.EscapeForRustDebug</c> keep compiling.
+        /// </summary>
+        internal static string EscapeForRustDebug(string s) => SearchIdentity.EscapeForRustDebug(s);
     }
 
     /// <summary>
